+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [35.92.0] - 2026-09-19
+
+### Added
+- Links to published task instructions in volunteer reminder emails, automatically matched to the task type.
+- Playing-day labels beside teams in volunteer statistics to distinguish Saturday and Sunday teams.
+
+### Fixed
+- Sort team names naturally by age group and team number in ascending and descending order.
+
 ## [35.91.0] - 2026-09-19
 
 ### Added